+++ v3/crates/engine/tests/introspection.rs
@@ -171,7 +171,20 @@ fn test_introspect_aggregates_root_field_nested_object() -> anyhow::Result<()> {
         test_path_string,
         &[
             "execute/aggregates/common_metadata/custom_connector_v02_schema.json",
-            "execute/aggregates/common_metadata/custom_connector_types.json",
+            "execute/aggregates/common_metadata/custom_connector_vBoth_types.json",
+            "execute/aggregates/common_metadata/supergraph.json",
+        ],
+    )
+}
+
+#[test]
+fn test_introspect_aggregates_root_field_custom_count_return_type() -> anyhow::Result<()> {
+    let test_path_string = "execute/aggregates/root_field/custom_count_return_type";
+    common::test_introspection_expectation(
+        test_path_string,
+        &[
+            "execute/aggregates/common_metadata/custom_connector_v02_schema.json",
+            "execute/aggregates/common_metadata/custom_connector_v02_types.json",
             "execute/aggregates/common_metadata/supergraph.json",
         ],
     )